Prevent new instance attributes from being filled
The problem: as the first step in `newFromBuilder`, `newInstance` is
being called with all of the model's attributes. `newInstance` then
passes the attributes into the Model constructor (via `getChildModel`).
The Model constructor then passes the attributes into `fill`. This means
that attribute mutators are called on model initialization, which is NOT
desirable.

Compare this to Eloquent's default `newFromBuilder` method, which passes
an empty array into `newInstance`. The model is initialized without
attributes, and then `setRawAttributes` is called to set the attributes
without instigating the attribute mutators.

Of course, it's important for `newInstance` to have the value of the
inheritance column attribute so that it can work out what class to
instantiate. This commit adjusts the `newFromBuilder` method so that
this is the only attribute passed into `newInstance`; the rest are then
set via `setRawAttributes`.